Fix improperly commented out macros in %%prep
Since RPM 4.20, %prep is no longer implemented as a special "hack" [1]
but is now a regular build scriptlet, meaning that %setup or %patch are
now treated as regular macros and thus will be expanded even in comments
(as documented in [2]).

Our spec has these so fix them up to unbreak the build on F41.

[1] https://github.com/rpm-software-management/rpm/issues/2205
[2] https://rpm-software-management.github.io/rpm/manual/spec.html